requirements.txt
fastapi
uvicorn
python-multipart
qdrant-client
openai
pdfplumber
pydantic
pymongo
google-cloud-storage
motor
python-dotenv
fastmcp
PyYAML
aiosqlite
